refactor(auth): centralize signin destination handling
