+++ apps/main/lib/presentation/modules/auth/signin/views/signin.action.dart
@@ -2,7 +2,6 @@ part of 'signin_screen.dart';
 
 extension SignInAction on SignInScreenState {
   void loginSuccessCallback() {
-    final destination = widget.redirectTo ?? DevModeDashboardScreen.routeName;
-    const PushReplacementNamedBehavior().push(context, destination);
+    context.completeSignIn(returnTo: widget.redirectTo);
   }
 }
